Submition for Level 10 in WD201

+++ app.js
@@ -1,38 +1,178 @@
 const express = require("express");
+const flash = require("connect-flash");
 var csrf = require("tiny-csrf")
 const app = express();
-const { Todo } = require("./models");
+const { Todo, User } = require("./models");
 const bodyParser = require("body-parser");
 var cookieParser = require("cookie-parser");
 
-// const { fileURLToPath } = require ("url"); // the node package 'url'
+const passport = require('passport');
+const connectEnsureLogin = require('connect-ensure-login');
+const session = require('express-session');
+const LocalStrategy = require('passport-local');
 
-// function dirname(meta) {
-//     return fileURLToPath(meta.url);
-// }
+const bcrypt = require('bcrypt');
+const saltRounds = 10;
 
-// // call with import.meta
-// const __dirname = dirname(meta);
+const path = require("path");
+//const __dirname = path.resolve();
+// eslint-disable-next-line no-undef
+app.set("views", path.join(__dirname, "views"));
+app.use(flash());
 
-//const path = require("path")
-//app.use(express.static(path.join(__dirname,"public")));
 
 
 app.use(bodyParser.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser("shh! some secret string"));
 app.use(csrf("123456789iamasecret987654321look", ["POST", "PUT", "DELETE"]));
-
 app.use(express.static("public"));
+
+app.use(session({
+  resave: false,//added 
+  saveUninitialized: true,//added 
+  secret: "my-super-secret-key-123123123123",
+  cookie: {
+    maxAge: 24 * 60 * 60 * 100
+  }
+}));
+
+app.use(passport.initialize());
+app.use(passport.session());
+
+passport.use(new LocalStrategy({
+  usernameField: 'email',
+  passwordField: 'password'
+}, (username, password, done) => {
+  console.log("Authentication")
+  User.findOne({ where: { email: username } })
+    .then(async (user) => {
+      console.log("Check password for " + user.username)
+      const result = await bcrypt.compare(password, user.password);
+      console.log(result)
+      if (result) {
+        return done(null, user);
+      }
+      else {
+        console.log("Invalid user");
+        return done(null, false, { message: "Invalid Username/Password" });
+      }
+
+    })
+    .catch((error) => {
+      return (error)
+    })
+}
+))
+
+passport.serializeUser((user, done) => {
+  console.log("Serializing user in session", user.id);
+  done(null, user.id);
+});
+
+passport.deserializeUser((id, done) => {
+  User.findByPk(id)
+    .then(user => {
+      done(null, user)
+    })
+    .catch((error) => {
+      done(error, null)
+    })
+})
 
 app.set("view engine", "ejs");
 
-app.get("/", async function (request, response) {
+app.use(function (request, response, next) {
+  response.locals.messages = request.flash();
+  next();
+});
+
+
+
+app.get("/home", (request, response) => {
+  response.render("home");
+});
+
+app.get("/signup", (request, response) => {
+  return response.render("signup", { csrfToken: request.csrfToken() });
+});
+
+app.get("/login", (request, response) => {
+  response.render("login", { csrfToken: request.csrfToken() });
+});
+
+app.post("/session",
+  passport.authenticate('local', {
+    failureRedirect: "/login",
+    failureFlash: true,
+  }),
+  (request, response) => {
+    response.redirect("/")
+  });
+
+app.get("/signout", (request, response, next) => {
+  request.logout((err) => {
+    if (err)
+      return next(err);
+    else
+      return response.redirect("/home");
+  })
+});
+app.post("/users", async (request, response) => {
+  const hashedPassword = await bcrypt.hash(request.body.password, saltRounds);
+  try {
+    console.log("Going to create new user")
+
+    //const newUser = 
+    User.create({
+      firstName: request.body.firstName,
+      lastName: request.body.lastName,
+      email: request.body.email,
+      password: hashedPassword
+    })
+      .then(user => {
+        request.login(user, (err) => {
+          if (err) {
+            console.log("Login Issue:" + err);
+          }
+        });
+        console.log(user);
+        response.redirect("/");
+      })
+      .catch(err => {
+        console.log("Error at Sign up:" + err);
+        //request.flash('error', err);
+        err.errors.map(e => {
+          console.log(e.message);
+          request.flash('error', { message: e.message });
+        });
+        
+        response.redirect("/signup");
+        //throw err;
+      });
+
+  } catch (error) {
+    console.log("Sign up Error at post:\n");
+
+    //request.flash('error', { message: error.messages });
+    //return response.redirect("/");
+    response.redirect("/signup");
+  }
+});
+
+app.get("/users", async (request, response) => {
+  const allUsers = await User.getUsers();
+  response.json({ allUsers });
+})
+
+app.get("/", connectEnsureLogin.ensureLoggedIn(), async function (request, response) {
+
+  const loggedUserId = request.user.id;
   const allTodos = await Todo.getTodos();
-  const overdueTodos = await Todo.getOverdueTodos();
-  const dueTodayTodos = await Todo.getDueTodayTodos();
-  const dueLaterTodos = await Todo.getDueLaterTodos();
-  const completedItems = await Todo.getCompletedItems();
+  const overdueTodos = await Todo.getOverdueTodos(loggedUserId);
+  const dueTodayTodos = await Todo.getDueTodayTodos(loggedUserId);
+  const dueLaterTodos = await Todo.getDueLaterTodos(loggedUserId);
+  const completedItems = await Todo.getCompletedItems(loggedUserId);
 
   if (request.accepts("html")) {
     response.render("index", {
@@ -47,10 +187,11 @@ app.get("/", async function (request, response) {
   else {
     response.json({ allTodos });
   }
+
 });
 
 
-app.get("/todos", async function (request, response) {
+app.get("/todos", connectEnsureLogin.ensureLoggedIn(), async function (request, response) {
   console.log("Processing list of all Todos ...");
   // FILL IN YOUR CODE HERE
 
@@ -63,7 +204,7 @@ app.get("/todos", async function (request, response) {
   response.send(todos)
 });
 
-app.get("/todos/:id", async function (request, response) {
+app.get("/todos/:id", connectEnsureLogin.ensureLoggedIn(), async function (request, response) {
   try {
     const todo = await Todo.findByPk(request.params.id);
     return response.json(todo);
@@ -73,15 +214,19 @@ app.get("/todos/:id", async function (request, response) {
   }
 });
 
-app.post("/todos", async function (request, response) {
+app.post("/todos", connectEnsureLogin.ensureLoggedIn(), async function (request, response) {
   // Create a todo
   try {
-    const todo = await Todo.addTodo(request.body);
+    const loggedUserId = request.user.id;
+    console.log("Logged User:::" + loggedUserId);
+    const todo = await Todo.addTodo(request.body, loggedUserId);
     // console.log("Accept:-" + request.accepts("json"));
     // console.log("Accept:-" + request.accepts("html"));
-
+    console.log("Inserted Todos");
     if (request.accepts("html")) {
       // console.log("HTML response:" + response.text)
+
+      console.log("Html Request");
       return response.redirect("/");
     }
     else {
@@ -90,26 +235,32 @@ app.post("/todos", async function (request, response) {
 
   } catch (error) {
     console.log("Error at post:" + error);
-    return response.status(422).json(error);
+    request.flash('error', { message: error.message.split(":")[1] });
+    return response.redirect("/");
   }
 });
 
 // app.put("/todos/:id/markAsCompleted", async function (request, response) {
-app.put("/todos/:id", async function (request, response) {
+app.put("/todos/:id", connectEnsureLogin.ensureLoggedIn(), async function (request, response) {
   console.log("Testing:::::::::" + request.params.id + " --- " + request.body.completed);
   const todo = await Todo.findByPk(request.params.id);
+  console.log("Todo:-" + todo.displayableString());
   try {
     //const updatedTodo = await todo.markAsCompleted();
-    const updatedTodo = await todo.setCompletionStatus(request.body.completed);
-    return response.json(updatedTodo);
-
+    if (todo.userId == request.user.id) {
+      const updatedTodo = await todo.setCompletionStatus(request.body.completed);
+      return response.json(updatedTodo);
+    }
+    else {
+      return {};
+    }
   } catch (error) {
     console.log("PUT Request:" + error);
     return response.status(422).json(error);
   }
 });
 
-app.delete("/todos/:id", async function (request, response) {
+app.delete("/todos/:id", connectEnsureLogin.ensureLoggedIn(), async function (request, response) {
   console.log("We have to delete a Todo with ID: ", request.params.id);
   // First, we have to query our database to delete a Todo by ID.
   const todo = await Todo.findByPk(request.params.id);
@@ -117,9 +268,13 @@ app.delete("/todos/:id", async function (request, response) {
     // console.log(todo);
     if (todo == null)
       return response.send(false);
-
-    todo.deleteTodo();
-    return response.send(true);
+    if (todo.userId == request.user.id) {
+      todo.deleteTodo();
+      return response.send(true);
+    }
+    else {
+      return response.send(false);
+    }
   } catch (error) {
     console.log(error);
     return response.status(422).json(error);
